requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
amqp==5.1.1
    # via kombu
anyio==4.6.0
    # via httpx
asgiref==3.7.2
    # via django
async-timeout==4.0.2
    # via redis
attrs==23.1.0
    # via
    #   jsonschema
    #   referencing
billiard==4.1.0
    # via celery
bleach[css]==5.0.1
    # via django-bleach
build==1.0.3
    # via pip-tools
celery==5.3.1
    # via
    #   django-celery-beat
    #   django-celery-email
    #   django-celery-results
    #   flower
certifi==2024.7.4
    # via
    #   -r requirements.in
    #   elasticsearch
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.2.0
    # via requests
click==8.1.6
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cortex4py==2.1.0
    # via -r requirements.in
cron-descriptor==1.4.0
    # via django-celery-beat
decorator==5.1.1
    # via validators
django==4.2.16
    # via
    #   -r requirements.in
    #   django-appconf
    #   django-auditlog
    #   django-bleach
    #   django-celery-beat
    #   django-celery-email
    #   django-comments-dab
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-jsonfield-backport
    #   django-lifecycle
    #   django-model-utils
    #   django-netfields
    #   django-nine
    #   django-picklefield
    #   django-timezone-field
    #   django-treebeard
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-appconf==1.0.5
    # via django-celery-email
django-auditlog==2.3.0
    # via -r requirements.in
django-bleach==3.0.0
    # via -r requirements.in
django-celery-beat==2.5.0
    # via -r requirements.in
django-celery-email==3.0.0
    # via -r requirements.in
django-celery-results==2.4.0
    # via -r requirements.in
django-colorfield==0.10.1
    # via -r requirements.in
django-comments-dab==2.7.1
    # via -r requirements.in
django-constance==3.1.0
    # via -r requirements.in
django-cors-headers==3.12.0
    # via -r requirements.in
django-debug-toolbar==3.4.0
    # via -r requirements.in
django-elasticsearch-dsl==7.3
    # via django-elasticsearch-dsl-drf
django-elasticsearch-dsl-drf==0.22.5
    # via -r requirements.in
django-extensions==3.1.5
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
django-jsonfield-backport==1.0.5
    # via -r requirements.in
django-lifecycle==1.0.0
    # via -r requirements.in
django-model-utils==4.2.0
    # via -r requirements.in
django-netfields==1.3.0
    # via -r requirements.in
django-nine==0.2.7
    # via django-elasticsearch-dsl-drf
django-picklefield==3.0.1
    # via
    #   -r requirements.in
    #   django-constance
django-timezone-field==5.1
    # via django-celery-beat
django-treebeard==4.5.1
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   django-elasticsearch-dsl-drf
    #   djangorestframework-simplejwt
    #   drf-spectacular
    #   rest-framework-generic-relations
djangorestframework-simplejwt==5.2.2
    # via -r requirements.in
drf-spectacular==0.26.4
    # via -r requirements.in
elasticsearch==7.17.9
    # via
    #   django-elasticsearch-dsl-drf
    #   elasticsearch-dsl
elasticsearch-dsl==7.4.1
    # via
    #   django-elasticsearch-dsl
    #   django-elasticsearch-dsl-drf
flower==2.0.1
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.6
    # via httpx
httpx==0.27.2
    # via whoisit
humanize==4.10.0
    # via flower
idna==3.7
    # via
    #   -r requirements.in
    #   anyio
    #   httpx
    #   requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.18.4
    # via drf-spectacular
jsonschema-specifications==2023.7.1
    # via jsonschema
kombu==5.3.1
    # via celery
netaddr==0.8.0
    # via django-netfields
packaging==23.1
    # via
    #   build
    #   django-lifecycle
    #   django-nine
    #   gunicorn
pillow==10.3.0
    # via
    #   -r requirements.in
    #   django-colorfield
pip-chill==1.0.1
    # via -r requirements.in
pip-tools==7.3.0
    # via -r requirements.in
prometheus-client==0.20.0
    # via flower
prompt-toolkit==3.0.39
    # via click-repl
psycopg2-binary==2.9.9
    # via -r requirements.in
publicsuffix2==2.20191221
    # via -r requirements.in
pyjwt==2.4.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
pyproject-hooks==1.0.0
    # via build
python-crontab==3.0.0
    # via django-celery-beat
python-dateutil==2.8.2
    # via
    #   celery
    #   django-auditlog
    #   elasticsearch-dsl
    #   python-crontab
    #   python-whois
    #   whoisit
python-magic==0.4.27
    # via cortex4py
python-whois==0.9.4
    # via -r requirements.in
pytz==2023.3
    # via
    #   django-timezone-field
    #   flower
pyyaml==6.0.1
    # via drf-spectacular
redis==4.4.4
    # via -r requirements.in
referencing==0.30.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   -r requirements.in
    #   cortex4py
    #   whoisit
rest-framework-generic-relations==2.1.0
    # via -r requirements.in
rpds-py==0.9.2
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via
    #   bleach
    #   django-elasticsearch-dsl
    #   django-elasticsearch-dsl-drf
    #   django-netfields
    #   elasticsearch-dsl
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlparse==0.5.1
    # via
    #   -r requirements.in
    #   django
    #   django-debug-toolbar
tinycss2==1.1.1
    # via
    #   -r requirements.in
    #   bleach
tld==0.12.6
    # via -r requirements.in
tornado==6.4.1
    # via flower
typing==3.7.4.3
    # via cortex4py
typing-extensions==4.12.2
    # via whoisit
tzdata==2023.3
    # via
    #   celery
    #   django-celery-beat
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.19
    # via
    #   -r requirements.in
    #   elasticsearch
    #   requests
urlman==2.0.1
    # via django-lifecycle
validators==0.19.0
    # via -r requirements.in
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
wheel==0.41.0
    # via pip-tools
whoisit==3.0.4
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
